TEST: Move variable tests into one class.
